+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ### Changelog
 
+### 1/15/2016
+
+* Fix: respond to @bot - [@dblock](https://github.com/dblock).
+
 ### 1/6/2016
 
 * [#1](https://github.com/dblock/slack-metabot/issues/1): Added JSON processing - [@dblock](https://github.com/dblock).